+++ src/main/scala/Day1/HelloWorld.scala
@@ -1,0 +1,18 @@
+package Day1
+
+object HelloWorld {
+
+  def helloWorldPrint(input: String): Unit = {
+    println(input)
+  }
+
+  def helloWorld(input: String): String = {
+    input
+  }
+
+  //TODO
+  def varibleTypeCheck(input: Unit): Unit = {
+    println(input)
+  }
+
+}
