+++ src/main/scala/Day1/ReduceString.scala
@@ -1,0 +1,8 @@
+package Day1
+
+object ReduceString{
+  def reduceString(string: String, int: Int): String = {
+    string.takeRight(int)
+  }
+
+}
